Fixed JSON serialization of dynamic assemblies
This change was made to combat flaky tests when running Rebus with
FakeItEasy and xUnit. xUnit runs unit tests in parallell and Newtonsoft
uses Assembly.LoadPartialWithName. For some reason, Castle.Core (which
FakeItEasy uses) sometimes creates a dynamic assembly with a public key
and sometimes it creates one without a public key. When two tests are
run in parallel, there can be two dynamic assemblies with the same name
that only differ in public key. LoadPartialWithName then gets confused
and throws an exception. This change tells Newtonsoft to use full
assembly names when serializing which circumvents the issue.